+++ HISTORY.md
@@ -1,6 +1,7 @@
 unreleased
 ==========
 
+  * Fix JSON strict violation error to match native parse error
   * deps: bytes@3.0.0
   * deps: debug@2.6.8
   * deps: depd@~1.1.1